+++ playground/nuxt.config.ts
@@ -3,10 +3,19 @@ export default defineNuxtConfig({
     '@nuxthq/ui',
     '../src/module'
   ],
+  // ssr: false,
   mdc: {
     highlight: {
-      theme: 'github-dark'
+      theme: {
+        default: 'vitesse-light',
+        dark: 'material-theme-palenight',
+      },
+      preload: [
+        'sql'
+      ]
     }
   },
-  devtools: { enabled: true }
+  devtools: {
+    enabled: true
+  }
 })