+++ src/module.ts
@@ -1,9 +1,11 @@
-import { defineNuxtModule, addComponent, addComponentsDir, createResolver, addServerHandler, addTemplate, addImports } from '@nuxt/kit'
+import { defineNuxtModule, extendViteConfig, addComponent, addComponentsDir, createResolver, addServerHandler, addTemplate, addImports } from '@nuxt/kit'
 import fs from 'fs'
 import { mdcImportTemplate } from './utils/templates'
 import type { ModuleOptions } from './types'
 import { defu } from 'defu'
 import { registerMDCSlotTransformer } from './utils/vue-mdc-slot'
+import { pathToFileURL } from 'url'
+import type { Theme } from './runtime/shiki/types'
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
@@ -17,9 +19,12 @@ export default defineNuxtModule<ModuleOptions>({
     highlight: false,
     headings: {
       anchorLinks: {
+        h1: false,
         h2: true,
         h3: true,
-        h4: true
+        h4: true,
+        h5: false,
+        h6: false
       }
     },
     components: {
@@ -37,6 +42,13 @@ export default defineNuxtModule<ModuleOptions>({
       },
       headings: options.headings!
     })
+    nuxt.options.runtimeConfig.mdc = defu(nuxt.options.runtimeConfig.mdc, {
+      highlight: options.highlight ? {
+        theme: options.highlight!.theme!,
+        preload: options.highlight!.preload!,
+        wrapperStyle: options.highlight!.wrapperStyle!
+      } : {}
+    })
 
     nuxt.hook('vite:extendConfig', (viteConfig) => {
       viteConfig.optimizeDeps?.include?.push(
@@ -45,7 +57,8 @@ export default defineNuxtModule<ModuleOptions>({
     })
 
     // Add imports template
-    nuxt.options.alias['#mdc-imports'] = addTemplate({ filename: 'mdc-imports.mjs', getContents: mdcImportTemplate, options, write: true }).dst
+    const { dst: templatePath } = addTemplate({ filename: 'mdc-imports.mjs', getContents: mdcImportTemplate, options, write: true })
+    nuxt.options.alias['#mdc-imports'] = process.env.NODE_ENV === 'development' ? pathToFileURL(templatePath).href : templatePath
     nuxt.options.nitro.alias = nuxt.options.nitro.alias || {}
     nuxt.options.nitro.alias['#mdc-imports'] = nuxt.options.alias['#mdc-imports']
 
@@ -67,8 +80,19 @@ export default defineNuxtModule<ModuleOptions>({
       })
     }
 
-    // Add server handlers
-    addServerHandler({ route: '/api/_mdc/highlight', handler: resolver.resolve('./runtime/shiki/event-handler') })
+    // Enable wasm for shikiji
+    if (options.highlight) {
+      nuxt.options.nitro.experimental = nuxt.options.nitro.experimental || {}
+      nuxt.options.nitro.experimental.wasm = true
+      // Add server handlers
+      addServerHandler({ route: '/api/_mdc/highlight', handler: resolver.resolve('./runtime/shiki/event-handler') })
+    }
+
+    extendViteConfig((config) => {
+      config.optimizeDeps = config.optimizeDeps || {}
+      config.optimizeDeps.exclude = config.optimizeDeps.exclude || []
+      config.optimizeDeps.exclude.push('@nuxtjs/mdc')
+    })
 
     // Register user global components
     const _layers = [...nuxt.options._layers].reverse()
@@ -77,7 +101,7 @@ export default defineNuxtModule<ModuleOptions>({
       const globalComponents = resolver.resolve(srcDir, 'components/mdc')
       const dirStat = await fs.promises.stat(globalComponents).catch(() => null)
       if (dirStat && dirStat.isDirectory()) {
-        nuxt.hook('components:dirs', (dirs) => {
+        nuxt.hook('components:dirs', (dirs: any[]) => {
           dirs.unshift({
             path: globalComponents,
             global: true,
@@ -93,6 +117,15 @@ export default defineNuxtModule<ModuleOptions>({
 })
 
 declare module '@nuxt/schema' {
+  interface RuntimeConfig {
+    mdc: {
+      highlight: {
+        theme?: Theme
+        preload?: string[]
+        wrapperStyle?: boolean | string
+      }
+    }
+  }
   interface PublicRuntimeConfig {
     mdc: {
       components: {