+++ src/runtime/utils/node.ts
@@ -32,7 +32,7 @@ export function isTag (vnode: VNode | MDCNode, tag: string | symbol): boolean {
  * Check if virtual node is text node
  */
 export function isText (vnode: VNode | MDCNode): boolean {
-  return isTag(vnode, 'text') || typeof (vnode as VNode).children === 'string'
+  return isTag(vnode, 'text') || isTag(vnode, Symbol.for('v-txt'))
 }
 
 /**
@@ -65,13 +65,13 @@ export function nodeTextContent (node: VNode | MDCNode): string {
   }
 
   if (isText(node)) {
-    return (node as VNode).children as string || (node as MDCText).value!
+    return (node as VNode).children as string || (node as MDCText).value! || ''
   }
 
   // Walk through node children
   const children = nodeChildren(node as MDCElement)
   if (Array.isArray(children)) {
-    return children.map(nodeTextContent).join('')
+    return children.map(nodeTextContent).filter(Boolean).join('')
   }
 
   // Return empty string for non-text nodes without any children
@@ -84,7 +84,7 @@ export function nodeTextContent (node: VNode | MDCNode): string {
  * @param tags list of tags to unwrap
  * @returns
  */
-export function unwrap (vnode: VNode, tags = ['p']): VNode | VNode[] {
+export function unwrap (vnode: VNode, tags: string[] = []): VNode | VNode[] {
   if (Array.isArray(vnode)) {
     return vnode.flatMap(node => unwrap(node, tags))
   }
@@ -102,7 +102,7 @@ export function unwrap (vnode: VNode, tags = ['p']): VNode | VNode[] {
   return result
 }
 
-function _flatUnwrap (vnodes: VNode | VNode[], tags = ['p']): Array<VNode> {
+function _flatUnwrap (vnodes: VNode | VNode[], tags: string[] = []): Array<VNode> {
   vnodes = Array.isArray(vnodes) ? vnodes : [vnodes]
 
   if (!tags.length) {
@@ -114,14 +114,22 @@ function _flatUnwrap (vnodes: VNode | VNode[], tags = ['p']): Array<VNode> {
     .filter(vnode => !(isText(vnode) && nodeTextContent(vnode).trim() === ''))
 }
 
-export function flatUnwrap (vnodes: VNode | VNode[], tags = ['p']): Array<VNode | string> {
-  return _flatUnwrap(vnodes, tags)
+export function flatUnwrap (vnodes: VNode | VNode[], tags: string | string[] = []): Array<VNode | string> | VNode {
+  if (typeof tags === 'string') {
+    tags = tags.split(',').map(tag => tag.trim()).filter(Boolean)
+  }
+
+  if (!tags.length) {
+    return vnodes
+  }
+
+  return _flatUnwrap(vnodes, tags as unknown as string[])
     .reduce((acc, item) => {
-      if (typeof item.children === 'string') {
+      if (isText(item)) {
         if (typeof acc[acc.length - 1] === 'string') {
-          acc[acc.length - 1] += item.children
+          acc[acc.length - 1] += item.children as string
         } else {
-          acc.push(item.children)
+          acc.push(item.children as string)
         }
       } else {
         acc.push(item)