+++ src/runtime/utils/ssrSlot.ts
@@ -5,6 +5,6 @@ export const ssrRenderSlot = (slots: Record<string, any>, name: string, props: a
   if (slots[name]) {
     return _ssrRenderSlot({ ...slots, [name]: () => flatUnwrap(slots[name](), props?.unwrap) }, name, props, fallbackRenderFn, push, parentComponent, slotScopeId)
   }
-  
+
   return _ssrRenderSlot(slots, name, props, fallbackRenderFn, push, parentComponent, slotScopeId)
 }