+++ src/utils/vue-mdc-slot.ts
@@ -1,31 +1,9 @@
-import { RENDER_SLOT, type NodeTransform, type ElementNode } from '@vue/compiler-core'
+import { type NodeTransform, type ElementNode } from '@vue/compiler-core'
 import { type Resolver, extendViteConfig } from '@nuxt/kit'
 
 export const registerMDCSlotTransformer = (resolver: Resolver) => {
     extendViteConfig((config) => {
     const compilerOptions = (config as any).vue.template.compilerOptions
-
-    config.plugins = config.plugins || []
-    config.plugins!.push({
-      name: 'mdc-render-slot-import',
-      enforce: 'post',
-      transform(code) {
-        if (code.includes('_renderMDCSlot')) {
-          return {
-            code: `import { renderSlot as _renderMDCSlot } from '${resolver.resolve('./runtime/utils/slot')}';\n${code}`,
-            map: { mappings: '' }
-          }
-        }
-
-        if (code.includes('_ssrRenderMDCSlot')) {
-          return {
-            code: `import { ssrRenderSlot as _ssrRenderMDCSlot } from '${resolver.resolve('./runtime/utils/ssrSlot')}';\n${code}`,
-            map: { mappings: '' }
-          }
-        }
-      }
-    })
-
     compilerOptions.nodeTransforms = [
       <NodeTransform>function viteMDCSlot(node: ElementNode, context) {
         if (node.tag === 'MDCSlot') {
@@ -41,7 +19,15 @@ export const registerMDCSlotTransformer = (resolver: Resolver) => {
             transform?.(node, context)
 
             const codegen = context.ssr ? (node as any).ssrCodegenNode : node.codegenNode
-            codegen.callee = codegen.callee === RENDER_SLOT ? '_renderMDCSlot' : '_ssrRenderMDCSlot'
+            codegen.callee = context.ssr ? '_ssrRenderMDCSlot' : '_renderMDCSlot'
+
+            const importExp = context.ssr ? '{ ssrRenderSlot as _ssrRenderMDCSlot }' : '{ renderSlot as _renderMDCSlot }'
+            if (!context.imports.some(i => String(i.exp) === importExp)) {
+              context.imports.push({
+                exp: importExp,
+                path: resolver.resolve(`./runtime/utils/${context.ssr ? 'ssrSlot' : 'slot'}`)
+              })
+            }
           } 
         }