+++ test/extras.test.ts
@@ -1,0 +1,6 @@
+import { expect, it } from 'vitest'
+import { Text } from 'vue'
+
+it('Vue.Text === Symbol.for(\'v-txt\')', async () => {
+  expect(Text === Symbol.for('v-txt')).toBe(true)
+})