+++ test/markdown/heading-id.test.ts
@@ -18,7 +18,7 @@ it('Heading id', async () => {
     const { data, body } = await parseMarkdown(md)
 
     expect(Object.keys(data)).toHaveLength(2)
-    
+
     expect(body.children).toMatchObject([
       { props: { id: 'hello'} },
       { props: { id: 'hello-world'} },